refactor: move new functions to CTypeLoader.ts
